Moved `#[derive(StableAbi)]` specific CommonTokens struct into the stable_abi submodule.
Declared structs in crate::common_tokens with the specific tokens used in each root module.
