+++ abi_stable_derive_lib/src/stable_abi/prefix_types.rs
@@ -30,6 +30,7 @@ use crate::{
 };
 
 use super::{
+    CommonTokens,
     attribute_parsing::{StabilityKind,StableAbiOptions},
     reflection::FieldAccessor,
 };
